+++ src/is_digits.c
@@ -3,7 +3,7 @@
 bool is_digits(char *str) {
     if(str == NULL) 
         return false;
-    for (int i = 0; i < strlen(str); i++)
+    for (size_t i = 0; i < strlen(str); i++)
     {
         if(!isdigit(str[i]))
             return false;
